feat: refactor extract function and delete redundant fields in database.
